+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
+## [1.5.4] - 2021-06-24
+- Improve the speed of auxillary script "rarefaction_curve.py"
+- Fix a bug in the Docker "run_instrain.py" script related to using compressed .fasta files
+
 ## [1.5.3] - 2021-03-26
 - BugFix related to inStrain clustering (https://github.com/MrOlm/inStrain/issues/52)
 - BugFix related to genes problems (https://github.com/MrOlm/inStrain/issues/53)